requirements.txt
aiosmtpd==1.4.4.post2
alembic==1.12.1
async-timeout==4.0.3
atpublic==4.0
attrs==23.1.0
Babel==2.14.0
blinker==1.7.0
certifi==2023.11.17
click==8.1.7
dnspython==2.4.2
dominate==2.9.0
elastic-transport==8.11.0
elasticsearch==8.11.1
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Admin==1.6.1
flask-babel==4.0.0
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.5.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.2
Pillow==10.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
rq==1.15.1
SQLAlchemy==2.0.23
typing_extensions==4.8.0
urllib3==2.1.0
visitor==0.1.3
Werkzeug==3.0.1
WTForms==3.1.1
